* Reversed order in dashboard * Fixed animations error after library update * WIP Adding new post page
