+++ package.json
@@ -26,7 +26,7 @@
     "angularfire2": "^4.0.0-rc.0",
     "core-js": "^2.4.1",
     "d3": "^4.5.0",
-    "firebase": "^3.9.0",
+    "firebase": "^4.1.3",
     "font-awesome": "^4.7.0",
     "hover.css": "^2.0.2",
     "jquery": "^3.1.1",
