+++ src/app/services/general.service.ts
@@ -25,6 +25,14 @@ export class GeneralService {
     return check;
   }
 
+  generateUID(): string {
+    let firstPart = ((Math.random() * 46656) | 0).toString (36);
+    let secondPart = ((Math.random() * 46656) | 0).toString (36);
+    firstPart = ("000" + firstPart).slice(-3);
+    secondPart = ("000" + secondPart).slice(-3);
+    return firstPart + secondPart;
+  }
+
   constructor (private auth: AngularFireAuth) {
     this.currentUser = this.auth.authState;
   }
