+++ src/styles/components/edit-post.component.css
