+++ src/styles/components/new-post.component.css
@@ -1,0 +1,12 @@
+:host
+{
+  width: 100%;
+  height: 100%;
+}
+
+.postref-img
+{
+  background: no-repeat center center;
+  background-size: cover;
+}
+
