+++ src/templates/components/login.component.html
@@ -24,6 +24,6 @@
     <div class="grid-35 mobile-portrait-grid-10 mobile-landscape-grid-30 tablet-portrait-grid-25 tablet-landscape-grid-25 h-50 mobile-portrait-h-50 mobile-landscape-h-50 tablet-portrait-h-50 tablet-landscape-h-50"></div>
   </div>
   <div class="w-100 h-20 mobile-portrait-h-10 mobile-landscape-h-10 tablet-portrait-h-10 tablet-landscape-h-20 v-center h-center">
-    <button type="submit" id="login-button" class="light-button h-50 mobile-portrait-h-80 mobile-landscape-h-100 tablet-portrait-h-70 tablet-landscape-h-50" [disabled]="!email || !validEmail() || !password"><span>Войти</span></button>
+    <button type="submit" id="login-button" class="light-button w-10 h-50 mobile-portrait-h-80 mobile-landscape-h-100 tablet-portrait-h-70 tablet-landscape-h-50" [disabled]="!email || !validEmail() || !password"><span>Войти</span></button>
   </div>
 </form>
