+++ src/templates/components/new-post.component.html
@@ -1,0 +1,56 @@
+<div class="grid-100 mobile-portrait-grid-100 mobile-landscape-grid-100 tablet-portrait-grid-100 tablet-landscape-grid-100 h-100">
+  <form (ngSubmit)="addPost()" id="content" class="scrollable-y w-100 h-100 mobile-portrait-h-100 mobile-landscape-h-100 tablet-portrait-h-100 tablet-landscape-h-100">
+    <div id="title-div" class="grid-100 mobile-portrait-grid-100 mobile-landscape-grid-100 tablet-portrait-grid-100 tablet-landscape-grid-100 h-10">
+      <div class="grid-30 h-100 v-center">
+        <label for="title-input" class="light-label">Название</label>
+      </div>
+      <div class="grid-30 h-100">
+        <input type="text" id="title-input" class="form-input-text inline-block w-100 h-60 mobile-portrait-h-40 mobile-landscape-h-50 tablet-portrait-h-50 tablet-landscape-h-60" name="title" [(ngModel)]="postRef.title">
+      </div>
+    </div>
+    <div id="subtitle-div" class="grid-100 mobile-portrait-grid-100 mobile-landscape-grid-100 tablet-portrait-grid-100 tablet-landscape-grid-100 h-10">
+      <div class="grid-30 h-100 v-center">
+        <label for="subtitle-input" class="light-label">Подзаголовок</label>
+      </div>
+      <div class="grid-30 h-100">
+        <input type="text" id="subtitle-input" class="form-input-text inline-block w-100 h-60 mobile-portrait-h-40 mobile-landscape-h-50 tablet-portrait-h-50 tablet-landscape-h-60" name="subtitle" [(ngModel)]="postRef.subtitle">
+      </div>
+    </div>
+    <div id="ref-div" class="grid-100 mobile-portrait-grid-100 mobile-landscape-grid-100 tablet-portrait-grid-100 tablet-landscape-grid-100 h-10">
+      <div class="grid-30 h-100 v-center">
+        <label for="ref-input" class="light-label">Ссылка</label>
+      </div>
+      <div class="grid-30 h-100">
+        <input type="text" id="ref-input" class="form-input-text inline-block w-100 h-60 mobile-portrait-h-40 mobile-landscape-h-50 tablet-portrait-h-50 tablet-landscape-h-60" name="ref" [(ngModel)]="postRef.ref">
+      </div>
+    </div>
+    <div id="img-div" class="grid-100 mobile-portrait-grid-100 mobile-landscape-grid-100 tablet-portrait-grid-100 tablet-landscape-grid-100 h-30">
+      <div class="grid-30 h-100 v-center">
+        <label for="img-input" class="light-label">Изображение</label>
+      </div>
+      <div class="grid-30 h-100 v-center">
+        <div class="w-50 h-100 v-center float-l">
+          <div class="postref-img w-100 h-80" [ngStyle]="{'background-image': 'url(' + postRef.img + ')' }"></div>
+        </div>
+        <div class="w-50 h-30 relative margin-left-10 pointer">
+          <button id="upload-button" class="light-button absolute w-100 h-100 no-padding" (click)="fireEvent ('img-input', 'click');">Выбрать</button>
+          <input type="file" id="img-input" class="inline-block w-100 h-100 pointer opacity-0" title=" " (change)="getImage ($event)">
+        </div>
+      </div>
+    </div>
+    <div id="text-div" class="grid-100 mobile-portrait-grid-100 mobile-landscape-grid-100 tablet-portrait-grid-100 tablet-landscape-grid-100 h-100">
+      <div class="grid-100 h-10 v-center">
+        <label for="text-input" class="light-label">Текст</label>
+      </div>
+      <textarea id="text-input"></textarea>
+    </div>
+    <div id="submit-div" class="grid-100 mobile-portrait-grid-100 mobile-landscape-grid-100 tablet-portrait-grid-100 tablet-landscape-grid-100 h-10">
+      <div class="grid-10 h-100">
+        <button type="submit" class="light-button grid-100 h-80">Добавить</button>
+      </div>
+      <div class="grid-10 h-100">
+        <button class="light-button grid-100 h-80" (click)="goBack()">Отмена</button>
+      </div>
+    </div>
+  </form>
+</div>
